refactor(ui): kebab-case 14 grandfathered filenames (warren-4e87)
Shrinks the `useFilenamingConvention` exception list in biome.jsonc from
30 entries to 16, taking the two groups that carry no risk of colliding
with work in flight: the seven `pages/run-analytics/*` charts, whose
only importer is RunAnalytics.tsx, and seven leaf components whose
importers are App.tsx, layout.tsx, Projects.tsx, Runs.tsx and
RunDetail.tsx.

Renames only. No file content changed beyond import specifiers, the
literal filenames the structural UI tests read, and the exception list
itself.

One entry moved rather than vanished: TokenConsumptionChart.tsx also
sits in the noExcessiveCognitiveComplexity exception block, so that
block now names token-consumption-chart.tsx. Dropping it instead of
renaming it would have failed the lint with a complexity error that
predates this change.
