+++ src/ui/src/components/auth-gate.tsx
