+++ src/ui/src/components/error-boundary.tsx
