+++ src/ui/src/components/warren-logo.tsx
