+++ src/ui/src/pages/RunDetail.tsx
@@ -6,8 +6,8 @@ import { projectsApi, runsApi } from "@/api/client.ts";
 import type { CancelRunResponse, PullRequestLifecycle } from "@/api/types.ts";
 import { isTerminalRunState } from "@/api/types.ts";
 import { OperatorOnly } from "@/components/OperatorOnly.tsx";
-import { StateBadge } from "@/components/StateBadge.tsx";
 import { RunFailureBadge } from "@/components/StatusIndicator.tsx";
+import { StateBadge } from "@/components/state-badge.tsx";
 import { Alert } from "@/components/ui/alert.tsx";
 import { Badge } from "@/components/ui/badge.tsx";
 import { Button } from "@/components/ui/button.tsx";
