+++ src/ui/src/pages/run-analytics/command-mining.tsx
