+++ src/ui/src/pages/run-analytics/token-consumption-chart.tsx
